chore: break monolithic format.rs into modules
